quickie for form validation in dao creation; probably should extend to spend and mint templates

+++ util/formValidation.ts
@@ -4,10 +4,10 @@ export const validateRequired = (v: string) =>
   v.length > 0 || 'Field is required'
 
 export const validatePositive = (v: string) =>
-  parseInt(v) > 0 || 'Must be positive'
+  parseFloat(v) > 0.0 || 'Must be positive'
 
 export const validateNonNegative = (v: string) =>
-  parseInt(v) >= 0 || 'Must be non-negative'
+  parseFloat(v) >= 0.0 || 'Must be non-negative'
 
 export const validatePercent = (v: string) => {
   const p = Number(v)
